requirements.txt
python-telegram-bot==20.0
requests
beautifulsoup4
gunicorn